# api/openapi/platform.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "星烽 StarBeacon 平台 API",
    "version": "0.1.0",
    "description": "仅描述仓库当前实现的 HTTP 接口。开发端通过 Vite 同源代理访问，生产使用 HTTPS 同源入口。64 位标识和字节计数以字符串传输；接口请求不接受自报租户。"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "SessionCookie": []
    }
  ],
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "summary": "登录",
        "description": "同源校验；每个连接来源 IP 每分钟最多 10 次。Cookie 为 HttpOnly、SameSite=Strict；生产启用 Secure。",
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Principal"
                }
              }
            }
          },
          "400": {
            "description": "请求无效、权限不足或依赖不可用，具体含义见 code/message。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "请求无效、权限不足或依赖不可用，具体含义见 code/message。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "请求无效、权限不足或依赖不可用，具体含义见 code/message。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "请求无效、权限不足或依赖不可用，具体含义见 code/message。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "请求无效、权限不足或依赖不可用，具体含义见 code/message。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "summary": "读取当前身份",
        "description": "",
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Principal"
                }
              }
            }
          },
          "400": {
            "description": "请求无效、权限不足或依赖不可用，具体含义见 code/message。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "请求无效、权限不足或依赖不可用，具体含义见 code/message。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "请求无效、权限不足或依赖不可用，具体含义见 code/message。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "请求无效、权限不足或依赖不可用，具体含义见 code/message。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "summary": "退出登录",
        "description": "",
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OK"
                }
              }
            }
          },
          "400": {
            "description": "请求无效、权限不足或依赖不可用，具体含义见 code/message。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "请求无效、权限不足或依赖不可用，具体含义见 code/message。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "请求无效、权限不足或依赖不可用，具体含义见 code/message。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "请求无效、权限不足或依赖不可用，具体含义见 code/message。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "description": "取自登录或 auth/me 返回的 csrf_token。",
            "schema": {
              "type": "string",
              "minLength": 64,
              "maxLength": 64
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "starbeacon_session"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "additionalProperties": false
      },
      "Login": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 64
          },
          "password": {
            "type": "string",
            "maxLength": 72
          }
        },
        "required": [
          "username",
          "password"
        ],
        "additionalProperties": false
      },
      "Principal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenant_id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "viewer"
            ]
          },
          "csrf_token": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "username",
          "role",
          "csrf_token"
        ],
        "additionalProperties": false
      },
      "OK": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "additionalProperties": false
      },
...
    }
  }
}